avoid filling stderr pipe & hanging server process
mininet sends the output of all .cmd(...)s to a pipe and reads the pipe
as part of the CLI loop. but if we send the process to the background,
mininet thinks it's completed so doesn't bother emptying the pipe. this
causes it to hange if it outputs enough to fill up the pipe. this commit
removes all logging from the python server, an MVP demonstrating a
solution to the issue.

See also:
https://github.com/mininet/mininet/issues/519#issuecomment-102198929